Implements the loading spinner and structure change
