+++ src/App.css
@@ -1,8 +1,23 @@
-.App-header {
-  min-height: 100vh;
+.header {
   display: flex;
   flex-direction: column;
   align-items: center;
   justify-content: center;
   color: black;
+}
+
+.main {
+  display: flex;
+  flex-direction: column;
+  border: 1px solid white;
+  border-radius: 8px;
+  padding: 20px;
+  margin-bottom: 20px;
+  gap: 8px;
+}
+
+.actions {
+  display: flex;
+  flex-direction: row;
+  gap: 8px;
 }
