+++ src/App.tsx
@@ -1,61 +1,54 @@
-import React, { useState } from "react";
+import { useState } from "react";
+import { createNewPatientsApi } from "./components/patients/PatientsApi";
+import { PatientsLoader } from "./components/patients/PatientsLoader";
+import { PatientsSearch } from "./components/patients/PatientsSearch";
+import { SearchSpinner } from "./components/spinner/SearchSpinner";
+import { Patient } from "./models/PatientModel";
+import { PatientService } from "./models/PatientServiceModel";
+import { PatientsDetail } from "./components/patients/PatientsDetail";
+import { ToastBox } from "./components/toast/ToastBox";
 import "./App.css";
-import { Patient, PatientsService } from "./patients/patients";
-import { createNewPatientsApi } from "./patients/patients_api";
-import { PatientsLoader } from "./patients/patients_loader_button";
-import { PatientsSearch } from "./patients/patients_search";
-import { ToastBox } from "./toast/toast";
+import "bootstrap/dist/css/bootstrap.min.css";
 
 function App() {
   const [patients, updatePatients] = useState<Patient[]>([]);
 
-  const [patientsApi] = useState<PatientsService>(
+  const [patientsApi] = useState<PatientService>(
     createNewPatientsApi("http://localhost:3000")
   );
 
+  const [isLoading, setLoading] = useState(false);
+
   return (
-    <div className="App">
-      <header className="App-header">
+    <>
+      <header className="header">
         <h1>Welcome to Heartbeat 🏥</h1>
-        <div
-          style={{
-            border: "1px solid white",
-            borderRadius: "8px",
-            padding: "20px",
-            marginBottom: "20px",
-          }}
-        >
-          <h2>Please load the patients using the button below or search</h2>
+      </header>
+      <div className="main">
+        <h4>Please load the patients using the button below or search</h4>
+        <div className="actions">
           <PatientsLoader
             loadPatients={patientsApi.All}
             onLoaded={updatePatients}
+            setLoading={setLoading}
+            disabled={isLoading}
           />
           <PatientsSearch
             loadPatients={patientsApi.Search}
             onResults={updatePatients}
+            setLoading={setLoading}
           />
-          {patients.length > 0 && displayPatients(patients)}
         </div>
+        { isLoading && <SearchSpinner /> }
+        { patients.length > 0 && <PatientsDetail patients={patients} /> }
         <ToastBox
           title="An error occurred"
           message="There was an error loading your results"
           status="success"
         />
-      </header>
-    </div>
+      </div>
+    </>
   );
 }
 
 export default App;
-
-function displayPatients(patients: Patient[]) {
-  return (
-    <ul>
-      {patients.map((p, k) => (
-        <li style={{ listStyle: "none" }} key={k}>
-          ✅ {p.name}{" "}
-        </li>
-      ))}
-    </ul>
-  );
-}
