+++ src/components/spinner/SearchSpinner.tsx
@@ -1,0 +1,8 @@
+import { FunctionComponent } from "react";
+import { Spinner } from "react-bootstrap";
+
+export const SearchSpinner: FunctionComponent = () => {
+    return (
+        <Spinner animation="border" variant="info" />
+    );
+}
