+++ src/components/toast/ToastBox.css
@@ -1,4 +1,4 @@
-.toast {
+/* .toast {
   min-width: 250px;
   margin-left: -125px;
   background-color: #333;
@@ -35,4 +35,4 @@
 .toast h6 {
   font-size: 1.2rem;
   margin: 0;
-}
+} */
