+++ src/models/PatientModel.tsx
@@ -1,0 +1,6 @@
+// Patient, is a patient in the Heartbeat system
+export type Patient = {
+    id: string;
+    name: string;
+    ehrID: string;
+};
