+++ src/models/PatientSearchQueryModel.tsx
@@ -1,0 +1,7 @@
+// PatientSearchQuery, a search query which allows
+// searching by name, id or ehrID.
+export type PatientSearchQuery = {
+    name?: string;
+    id?: string;
+    ehrID?: string;
+};
