+++ src/models/PatientServiceModel.tsx
@@ -1,0 +1,8 @@
+import { Patient } from "./PatientModel";
+import { PatientSearchQuery } from "./PatientSearchQueryModel";
+
+export interface PatientService {
+    All(): Promise<Patient[]>;
+    Get(id: string): Promise<Patient>;
+    Search(query: PatientSearchQuery): Promise<Patient[]>;
+}
